TP2/requirements.txt
diffusers>=0.27.0
transformers
accelerate
safetensors
torch
pillow
streamlit
